refac daemon module
RevaultD struct is a Client,
it is moved to the client module and
the jsonrpc client is moved to its own file.
